+++ src/Createable.java
@@ -2,5 +2,5 @@
  * Created by RikkiTheTramp on 14/05/18.
  */
 public interface Createable {
-    Chunk create(byte[] id, byte[] size);
+    Chunk create(int[] id, int[] size);
 }
