+++ src/JunkChunk.java
@@ -8,7 +8,7 @@ public class JunkChunk extends Chunk {
     }
 
     @Override
-    public void extractData(byte[] data) {
+    public void extractData(int[] data) {
         //DO NOTHING
     }
